requirements.txt
pyTelegramBotAPI
schedule
flask
